% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2002.R
\docType{data}
\name{io2002}
\alias{io2002}
\title{Input Output Tables at 2002}
\format{
An object of class \code{list} of length 30.
}
\usage{
io2002
}
\description{
A list that includes 31 provinces' IO tables at 2002.
}
\keyword{datasets}
