% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2007.R
\docType{data}
\name{io2007}
\alias{io2007}
\title{Input Output Tables at 2007}
\format{
An object of class \code{list} of length 30.
}
\usage{
io2007
}
\description{
A list that includes 31 provinces' IO tables at 2007.
}
\keyword{datasets}
