% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2012.R
\docType{data}
\name{io2012}
\alias{io2012}
\title{Input Output Tables at 2012}
\format{
An object of class \code{list} of length 31.
}
\usage{
io2012
}
\description{
A list that includes 31 provinces' IO tables at 2012.
}
\keyword{datasets}
