% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2017.R
\docType{data}
\name{io2017}
\alias{io2017}
\title{Input Output Tables at 2017}
\format{
An object of class \code{list} of length 31.
}
\usage{
io2017
}
\description{
A list that includes 31 provinces' IO tables at 2017.
}
\keyword{datasets}
