% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iodata.R
\docType{data}
\name{io2002}
\alias{io2002}
\title{Input Output Tables}
\format{
An object of class \code{list} of length 30.
}
\usage{
io2002
}
\description{
A dataset containing 4 lists that are io2002, io2007, io2012, io2017.
Every list include 31 provinces' IO tables .
}
\keyword{datasets}
